2D marker detection added
works with on marker of ID=0 showing the target position. If this is found the other markers are not used.
Otherwise, for each of the three planes there are two markers on a line passing through the center (id: 1,2, and 3 )
so that the target can be found as the intersection of any of these pair of lines.